refactor: IToken을 types으로 이동
데이터베이스 모델이 아닌 리턴타입으로 사용되므로 types 폴더로 이동
