--- src/auth/auth.model.ts
@@ -1,14 +1,0 @@
-export interface IToken {
-  uuid?: string;
-  TOKEN?: string;
-  REFRESH_TOKEN?: string;
-  verify?: string;
-}
-
-export default class implements IToken {
-  uuid?: string;
-
-  TOKEN?: string;
-
-  REFRESH_TOKEN?: string;
-}
