+++ src/shared/errors/auth.error.ts
@@ -1,6 +1,6 @@
 export const AuthError = {
   INVALID_TOKEN: {
-    code: 'AUTH-401',
+    code: 'auth-401',
     message: '유효하지 않은 토큰입니다.'
   }
 };
