+++ src/types/index.ts
@@ -3,3 +3,10 @@ interface ErrorDetail {
   message: string;
 }
 export type ErrorInfo = Record<string, ErrorDetail>;
+
+export interface IToken {
+  uuid?: string;
+  TOKEN?: string;
+  REFRESH_TOKEN?: string;
+  verify?: string;
+}
